Allow top level blocks for simpler testing
Use `accept` in parser visit functions, not `visit`.
Currently aggressively pruning the parse tree to generate a minimalistic AST tree. Might be flattening away too many nodes, if you think it's better to have a block that encapsulates a certain sequence of statements, it should be quite simple to add it back in the parser.
compilationUnit is still in the entry point, compoundStatement is added as a rule to compilationUnit to allow top level blocks. Top level expressions and statements are still not allowed, just wrap them in `{}`. This is to differentiate proper programs that we aim to parse in the end from our testing programs.