Change statements and declarations to not produce values
Add `main()` and return statements to most test cases. Should not need to use blocks as our entry point anymore.
Fix `break_i` that was broken because of a logical error in the while loop of the microcode.
Change statements and declarations to not produce values. The responsibility of popping the value off the OS if applicable and if necessary should be on the smallest enclosing node.

Change `value_assmt_i` to never pop the value. It was used for both declarations and assignment expressions. Now, declarations (ValueDeclaration) will push a POP_INSTRUCTION after the assignment is complete.
Ensure correctness by throwing an error with `pop_i` if the OS is empty.